+++ src/app/components/Lecture/Lecture.module.css
@@ -4,12 +4,6 @@
   width: 90vw;
 }
 
-.author-profile-img {
-  width: 32px;
-  height: 32px;
-  border-radius: 32px;
-}
-
 .profile {
   display: flex;
   justify-content: space-between;
@@ -19,6 +13,12 @@
   display: flex;
   align-items: center;
   margin-bottom: 10px;
+}
+
+.author-profile-img {
+  width: 32px;
+  height: 32px;
+  border-radius: 32px;
 }
 
 .author-box {
@@ -38,7 +38,7 @@
   display: flex;
 }
 
-.btn {
+.info button {
   color: #5cb85c;
 }
