--- src/app/pages/ConsultingDetail/ConsultingDetail.css
@@ -1,3 +1,0 @@
-.detail {
-  margin-top: 32px;
-}
