+++ src/app/pages/LectureDetail/LectureDetail.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
-import LECTURE_DETAIL_DATA from "../data/lectureDetailData";
+import LECTURE_DETAIL_DATA from "../../data/lectureDetailData";
 import { Link } from "react-router-dom";
+import css from "./LectureDetail.module.css";
 
 const LectureDetail = () => {
   const [data, setData] = useState([]);
@@ -14,15 +15,15 @@ const LectureDetail = () => {
       <div className="banner">
         <div className="container">
           <h1>{title}</h1>
-          <div className="profile-box detail">
+          <div className={css["profile-box"]}>
             <Link to={`/profile/${authorId}`}>
               <img
-                className="author-profile-img"
+                className={css["author-profile-img"]}
                 src={authorImage}
                 alt="강사 프로필 사진"
               />
             </Link>
-            <div className="author-box">
+            <div className={css["author-box"]}>
               <Link className="author" to={`/profile/${authorId}`}>
                 {authorName}
               </Link>
